Harden the INIT directory determination.
Original version ('..') expected your working directory to be where
the script was.  So calling any of the scripts from other locations
(e.g. 'run-parts ~/cloud/scripts') would have caused INIT variable
to be set relative to *your* location (not *the script's* one),
and thus not cause all sorts of problems for every next stage that
relies on it.

Using the '$(dirname "$BASH_SOURCE")/..' construct ties things relative
to the place where currently executed script is...  which is exactly
what we need.  More details in Wooledge FAQ at
http://mywiki.wooledge.org/BashFAQ/028